+++ app/mailers/user_mailer.rb
@@ -6,6 +6,7 @@ class UserMailer < ActionMailer::Base
     setup_email
     @system_admin = system_admin
     @user = user
+    @email_to = system_admin.email
     mail(to: system_admin.email,
          subject: "#{user.name} Signed Up",
          reply_to: user.email)
@@ -15,6 +16,7 @@ class UserMailer < ActionMailer::Base
     setup_email
     @system_admin = system_admin
     @agreement = agreement
+    @email_to = system_admin.email
     mail(to: system_admin.email,
          subject: "#{agreement.user.name} Submitted a Data Access and Use Agreement")
   end
@@ -22,6 +24,7 @@ class UserMailer < ActionMailer::Base
   def daua_approved(agreement, admin)
     setup_email
     @agreement = agreement
+    @email_to = agreement.user.email
     mail(to: agreement.user.email,
          subject: "Your DAUA Submission has been Approved",
          reply_to: admin.email)
@@ -29,7 +32,6 @@ class UserMailer < ActionMailer::Base
 
   def daua_progress_notification(agreement, admin)
     setup_email
-    attachments.inline['nsrr-logo.png'] = File.read('app/assets/images/nsrr_logo_64.png')
     @agreement = agreement
     @admin = admin
     @email_to = admin.email
@@ -43,6 +45,7 @@ class UserMailer < ActionMailer::Base
   def sent_back_for_resubmission(agreement, admin)
     setup_email
     @agreement = agreement
+    @email_to = agreement.user.email
     mail(to: agreement.user.email,
          subject: "Please Resubmit your DAUA",
          reply_to: admin.email)
@@ -50,19 +53,17 @@ class UserMailer < ActionMailer::Base
 
   def dataset_access_requested(dataset_user, editor)
     setup_email
-    attachments.inline['nsrr-logo.png'] = File.read('app/assets/images/nsrr_logo_64.png')
     @dataset_user = dataset_user
     @editor = editor
     @email_to = editor.email
     mail(to: @email_to,
-         subject: "#{dataset_user.user.name} Has Requested Dataset File Access on #{dataset_user.dataset.name}")
+         subject: "#{dataset_user.user.name} Requested Dataset File Access on #{dataset_user.dataset.name}")
   end
 
   protected
 
   def setup_email
-    # @footer_html = "<div style=\"color:#777\">Change #{DEFAULT_APP_NAME} email settings here: <a href=\"#{SITE_URL}/settings\">#{SITE_URL}/settings</a></div><br /><br />".html_safe
-    # @footer_txt = "Change #{DEFAULT_APP_NAME} email settings here: #{SITE_URL}/settings"
+    attachments.inline['nsrr-logo.png'] = File.read('app/assets/images/nsrr_logo_64.png')
   end
 
 end